// taurus-protect-sdk-typescript/src/internal/openapi/apis/GroupsApi.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Taurus-PROTECT APIs
 * Taurus-PROTECT APIs description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import * as runtime from '../runtime';
import type {
  GooglerpcStatus,
  TgvalidatordGetGroupsReply,
} from '../models/index';
import {
    GooglerpcStatusFromJSON,
    GooglerpcStatusToJSON,
    TgvalidatordGetGroupsReplyFromJSON,
    TgvalidatordGetGroupsReplyToJSON,
} from '../models/index';

export interface UserServiceGetGroupsRequest {
    limit?: string;
    offset?: string;
    ids?: Array<string>;
    externalGroupIds?: Array<string>;
    query?: string;
}

/**
 * GroupsApi - interface
 * 
 * @export
 * @interface GroupsApiInterface
 */
export interface GroupsApiInterface {
    /**
     * This endpoint returns a list of groups
     * @summary List groups
     * @param {string} [limit] 
     * @param {string} [offset] 
     * @param {Array<string>} [ids] 
     * @param {Array<string>} [externalGroupIds] 
     * @param {string} [query] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof GroupsApiInterface
     */
    userServiceGetGroupsRaw(requestParameters: UserServiceGetGroupsRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<TgvalidatordGetGroupsReply>>;

    /**
     * This endpoint returns a list of groups
     * List groups
     */
    userServiceGetGroups(requestParameters: UserServiceGetGroupsRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<TgvalidatordGetGroupsReply>;

}

/**
 * 
 */
export class GroupsApi extends runtime.BaseAPI implements GroupsApiInterface {

    /**
     * This endpoint returns a list of groups
     * List groups
     */
    async userServiceGetGroupsRaw(requestParameters: UserServiceGetGroupsRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<TgvalidatordGetGroupsReply>> {
        const queryParameters: any = {};

        if (requestParameters['limit'] != null) {
            queryParameters['limit'] = requestParameters['limit'];
        }

        if (requestParameters['offset'] != null) {
            queryParameters['offset'] = requestParameters['offset'];
        }

        if (requestParameters['ids'] != null) {
            queryParameters['ids'] = requestParameters['ids'];
        }

        if (requestParameters['externalGroupIds'] != null) {
            queryParameters['externalGroupIds'] = requestParameters['externalGroupIds'];
        }

        if (requestParameters['query'] != null) {
            queryParameters['query'] = requestParameters['query'];
        }

        const headerParameters: runtime.HTTPHeaders = {};

        if (this.configuration && this.configuration.apiKey) {
            headerParameters["Authorization"] = await this.configuration.apiKey("Authorization"); // ApiKeyTPV1 authentication
        }

        const response = await this.request({
            path: `/api/rest/v1/groups`,
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        }, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => TgvalidatordGetGroupsReplyFromJSON(jsonValue));
    }

    /**
     * This endpoint returns a list of groups
     * List groups
     */
    async userServiceGetGroups(requestParameters: UserServiceGetGroupsRequest = {}, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<TgvalidatordGetGroupsReply> {
        const response = await this.userServiceGetGroupsRaw(requestParameters, initOverrides);
        return await response.value();
    }

}